--- src/mocks/offers.js
@@ -1,90 +1,0 @@
-export const mockOffers = [
-  {
-    id: 1,
-    name: `Paris`,
-    title: `Beautiful & luxurious apartment at great location`,
-    price: 120,
-    type: `apartment`,
-    premium: true,
-    img: `img/apartment-01.jpg`,
-    rating: 93,
-    coordinates: [52.3909553943508, 4.85309666406198]
-  },
-  {
-    id: 2,
-    name: `Cologne`,
-    title: `Wood and stone place`,
-    price: 80,
-    type: `Private room`,
-    premium: false,
-    img: `img/room.jpg`,
-    rating: 80,
-    coordinates: [52.369553943508, 4.85309666406198]
-  },
-  {
-    id: 3,
-    name: `Brussels`,
-    title: `Canal View Prinsengracht`,
-    price: 132,
-    type: `apartment`,
-    premium: true,
-    img: `img/apartment-02.jpg`,
-    rating: 80,
-    coordinates: [52.3909553443508, 4.929309666406198]
-  },
-  {
-    id: 4,
-    name: `Amsterdam`,
-    title: `Nice, cozy, warm big bed apartment`,
-    price: 180,
-    type: `apartment`,
-    premium: true,
-    img: `img/apartment-03.jpg`,
-    rating: 100,
-    coordinates: [52.3509553943508, 4.939309666406198]
-  },
-  {
-    id: 5,
-    name: `Amsterdam`,
-    title: `Nice, cozy, warm big bed apartment`,
-    price: 180,
-    type: `apartment`,
-    premium: false,
-    img: `img/apartment-03.jpg`,
-    rating: 100,
-    coordinates: [52.3809355943508, 4.939309666406198]
-  },
-  {
-    id: 6,
-    name: `Hamburg`,
-    title: `Nice, cozy, warm big bed apartment`,
-    price: 180,
-    type: `apartment`,
-    premium: false,
-    img: `img/apartment-03.jpg`,
-    rating: 100,
-    coordinates: [52.3802345943508, 4.939309666406198]
-  },
-  {
-    id: 7,
-    name: `Hamburg`,
-    title: `Nice, cozy, warm big bed apartment`,
-    price: 180,
-    type: `apartment`,
-    premium: true,
-    img: `img/apartment-03.jpg`,
-    rating: 100,
-    coordinates: [52.3802377943508, 4.939309666406198]
-  },
-  {
-    id: 8,
-    name: `Dusseldorf`,
-    title: `Nice, cozy, warm big bed apartment`,
-    price: 180,
-    type: `apartment`,
-    premium: true,
-    img: `img/apartment-03.jpg`,
-    rating: 100,
-    coordinates: [52.3802377943000, 4.939309666406198]
-  },
-];
